+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Thanks to the following contributors who worked on this release:
 
 ### Fixed
 - Fixed a bug where duplicate submenus could be produced by add-ins with effect categories that were not translated (#1933, #1935)
+- The layers and history panels now scroll automatically to the selected item after any updates (#1867, #1828)
 
 ## [3.1.1](https://github.com/PintaProject/Pinta/release/tag/3.1.1) - 2026/01/10